Migraciones normales

+++ database/migrations/2021_11_19_115811_point_has_resources.php
@@ -15,8 +15,8 @@ class PointHasResources extends Migration
     {
         Schema::create('point_has_resources', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('idPoint');
-            $table->unsignedBigInteger('idResource');
+            $table->integer('idPoint');
+            $table->integer('idResource');
         });
     }
